requirements.txt
asgiref==3.7.2
cloudinary==1.37.0
Django==4.2.8
django-cloudinary-storage==0.3.0
django-etc==1.4.0
django-hitcount==1.3.5
django-money==3.4.1
django-taggit==5.0.1
djangorestframework==3.14.0
Pillow==10.1.0
py-moneyed==3.0
pytz==2023.3.post1
sqlparse==0.4.4
